requirements.txt
fastapi~=0.111
uvicorn~=0.29.0
streamlit~=1.33
browser-use
allure-pytest~=2.13
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
redis>=5.0,<6.0
networkx~=3.2.1
pytest~=8.0
pytest-bdd~=7.0
ruff>=0.1,<0.2
black>=24.0,<25.0
mypy>=1.0,<2.0
python-dotenv>=1.0,<2.0
requests~=2.31.0
types-requests~=2.31.0